[refactor] ErrorProps 분리
