+++ app/_components/providers/SyncThemeScript.tsx
