+++ app/_types/errorTypes.ts
@@ -1,0 +1,4 @@
+export interface ErrorProps {
+  error: Error & { digest?: string };
+  reset: () => void;
+}
